rename test provider package
