+++ cluster-autoscaler/cloudprovider/test/test_cloud_provider.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package testprovider
+package test
 
 import (
 	"fmt"
